% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{beta_data}
\alias{beta_data}
\title{Synthetic Beta data}
\format{A list with 300 elements, where each element element is an L x 2
  matrix of observations, where: \describe{ \item{1st column}{locations of
  obseravtions} \item{2nd column}{methylation level } }}
\usage{
beta_data
}
\value{
Synthetic Beta methylation data.
}
\description{
A synthetic dataset containinig 300 entries from Beta probit
  regression observations (e.g. Beta values from array methylation data).
}
\section{Details}{
 The beta regression model is based on alternative
  parameterization of the beta density in terms of the mean and dispersion
  parameter:\url{https://cran.r-project.org/web/packages/betareg/}.
}

\seealso{
\code{\link{binomial_data}}, \code{\link{bernoulli_data}},
  \code{\link{gaussian_data}}
}
\keyword{datasets}
