% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{binomial_data}
\alias{binomial_data}
\title{Synthetic Binomial data}
\format{
A list with 300 elements, where each element element is an L x 3
  matrix of observations, where: \describe{ \item{1st column}{locations of
  obseravtions} \item{2nd column}{total trials} \item{3rd column}{number of
  successes} }
}
\usage{
binomial_data
}
\value{
Synthetic Binomial methylation data.
}
\description{
A synthetic dataset containinig 300 entries from Binomial probit
  regression observations (e.g. bulk methylation data).
}
\seealso{
\code{\link{bernoulli_data}}, \code{\link{beta_data}},
  \code{\link{gaussian_data}}
}
\keyword{datasets}
