% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{boxplot_cluster_gex}
\alias{boxplot_cluster_gex}
\title{Boxplot of clustered expression levels}
\usage{
boxplot_cluster_gex(bpr_cluster_obj, gex, main_lab = "Gene expression levels")
}
\arguments{
\item{bpr_cluster_obj}{The output of the \code{bpr_cluster_wrap} function.}

\item{gex}{The vector of gene expression data for each promoter region.}

\item{main_lab}{The title of the plot}
}
\value{
The figure to be plotted in the device.
}
\description{
\code{boxplot_cluster_gex} creates a boxplot of clustered gene expression
levels which depend on the clustered methylation profiles. Each colour
denotes a different cluster.
}
\examples{
# Cluster methylation profiles using 4 RBFs
obs <- meth_data
basis <- create_rbf_object(M = 4)
res   <- bpr_cluster_wrap(x = obs, K = 3, em_max_iter = 5, opt_itnmax = 4,
                          init_opt_itnmax = 5, is_parallel = FALSE)

# Create the plot
boxplot_cluster_gex(bpr_cluster_obj = res, gex = gex_data)

}
\author{
C.A.Kapourani \email{C.A.Kapourani@ed.ac.uk}
}
\seealso{
\code{\link{plot_cluster_prof}}, \code{\link{plot_scatter_gex}},
  \code{\link{plot_fitted_profiles}}
}

