% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpr.R
\name{bpr_likelihood}
\alias{bpr_likelihood}
\title{BPR log likelihood function}
\usage{
bpr_likelihood(w, H, data, is_NLL = FALSE)
}
\arguments{
\item{w}{A vector of parameters (i.e. coefficients of the basis functions)}

\item{H}{The \code{L x M} matrix design matrix, where L is the number
of observations and M the number of basis functions.}

\item{data}{An \code{L x 2} matrix containing in the 1st column the total
number of trials and in the 2nd the number of successes. Each row
corresponds to each row of the design matrix.}

\item{is_NLL}{Logical, indicating if the Negative Log Likelihood should be
returned.}
}
\value{
the log likelihood
}
\description{
\code{bpr_likelihood} evaluates the Binomial distributed Probit regression
 log likelihood function for a given set of coefficients, observations and
 a design matrix.
}
\section{Mathematical formula}{

 The Binomial distributed Probit regression log likelihood function
 is computed by the following formula:
 \deqn{log p(y | f, w) = \sum_{l=1}^{L} log Binom(m_{l} | t_{l}, \Phi(w^{t}h(x_{l})))}
 where h(x_{l}) are the basis functions.
}
\examples{
obj <- polynomial.object(M=2)
obs <- c(0,.2,.5, 0.6)
des_mat <- design_matrix(obj, obs)
H <- des_mat$H
w <- c(.1,.1,.1)
data <- matrix(c(10,12,15,7,9,8), ncol=2)
lik <- bpr_likelihood(w, H, data)

}
\seealso{
\code{\link{bpr_gradient}}, \code{\link{design_matrix}}
}

