% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpr_optimize.R
\name{bpr_optim.matrix}
\alias{bpr_optim.matrix}
\title{Optimization method for the BPR NLL using matrix x}
\usage{
\method{bpr_optim}{matrix}(x, w = NULL, basis = NULL, fit_feature = NULL,
  cpg_dens_feat = FALSE, opt_method = "CG", opt_itnmax = 100, ...)
}
\arguments{
\item{x}{An L x 3 matrix of observations, where 1st column contains the
locations. The 2nd and 3rd columns contain the total trials and number of
successes at the corresponding locations, repsectively.}

\item{w}{A vector of parameters (i.e. coefficients of the basis functions)}

\item{basis}{A 'basis' object. E.g. see \code{\link{create_rbf_object}}.}

\item{fit_feature}{Additional feature on how well the profile fits the
methylation data.}

\item{cpg_dens_feat}{Additional feature for the CpG density across the
promoter region.}

\item{opt_method}{The optimization method to be used. See
\code{\link[stats]{optim}} for possible methods. Default is 'CG'.}

\item{opt_itnmax}{Optional argument giving the maximum number of iterations
for the corresponding method. See \code{\link[stats]{optim}} for details.}

\item{...}{Additional parameters}
}
\value{
A list containing the following elements:
\itemize{
 \item{ \code{w_opt}: Optimized values for the coefficient vector w.
   The length of the result is the same as the length of the vector w.
 }
 \item{ \code{basis}: The basis object.
 }
}
}
\description{
\code{bpr_optim.matrix} minimizes the negative log likelihood of the BPR
function. Since it cannot be evaluated analytically, an optimization
procedure is used. The \code{\link[stats]{optim}} packages is used for
performing optimization.
}
\examples{
basis <- create_rbf_object(M=2)
w <- c(0.1, 0.1, 0.1)
data <- meth_data[[1]]
out_opt <- bpr_optim(x = data, w = w, basis = basis, fit_feature = "NLL")

}
\author{
C.A.Kapourani \email{C.A.Kapourani@ed.ac.uk}
}
\seealso{
\code{\link{bpr_optim}}, \code{\link{bpr_optim.list}}
}

