% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpr_optimize.R
\name{bpr_optim}
\alias{bpr_optim}
\title{Generic function for optimizing BPR negative log likelihood function}
\usage{
bpr_optim(x, ...)
}
\arguments{
\item{x}{The input object}

\item{...}{Additional parameters}
}
\description{
\code{bpr_optim} is a generic function which calls the appropriate methods
depending on the class of the object \code{x}. Object \code{x} can be either
a \code{\link{list}} or a \code{\link{matrix}}.
}
\examples{
data <- meth_data
out_opt <- bpr_optim(x = data, is_parallel = FALSE, method = "BFGS")

}
\author{
C.A.Kapourani \email{C.A.Kapourani@ed.ac.uk}
}
\seealso{
\code{\link{bpr_optim.list}}, \code{\link{bpr_optim.matrix}}
}

