% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BPRMeth_package.R
\docType{package}
\name{BPRMeth}
\alias{.datatable.aware}
\alias{BPRMeth}
\alias{BPRMeth-package}
\title{\code{BPRMeth}: Extracting higher order methylation features}
\format{An object of class \code{logical} of length 1.}
\usage{
.datatable.aware
}
\description{
Higher order methylation features for clustering and prediction in epigenomic
studies
}
\author{
C.A.Kapourani \email{C.A.Kapourani@ed.ac.uk}
}
\keyword{datasets}

