% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_errors}
\alias{calculate_errors}
\title{Calculate error metrics}
\usage{
calculate_errors(x, y, summary = FALSE)
}
\arguments{
\item{x}{Actual values.}

\item{y}{Predicted values.}

\item{summary}{Logical, indicating if the erorr metrics should be printed.}
}
\value{
A list containing the following components:
\itemize{
 \item \code{mae} mean absolute error.
 \item \code{mse}  mean squared error (the variance).
 \item \code{rmse} root mean squared error (std dev).
 \item \code{mape} mean absolute percentage error.
 \item \code{rstd} relative standard deviation.
}
}
\description{
\code{calculate_errors} calculates error metrics so as to assess the
 performance of a model, e.g. linear regression.
}
\examples{
# Example data
actual <- c(4, 6, 9, 10, 4, 6, 4, 7, 8, 7)
predicted <- c(5, 6, 8, 10, 4, 8, 4, 9, 8, 9)

calculate_errors(actual, predicted, summary = TRUE)

}

