% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_methyl_region.R
\name{center_loc}
\alias{center_loc}
\title{Center CpG locations relative to TSS}
\usage{
center_loc(region, tss, strand_direction)
}
\arguments{
\item{region}{CpG locations}

\item{tss}{TSS location}

\item{strand_direction}{Strand direction}
}
\value{
Centered location data relative to TSS
}
\description{
\code{center_loc} centera CpG locations relative to TSS
}

