% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis_functions.R
\name{create_basis}
\alias{basis}
\alias{create_basis}
\alias{create_polynomial_object}
\alias{create_rbf_object}
\title{Create basis objects}
\usage{
create_rbf_object(M = 2, gamma = NULL, mus = NULL, eq_spaced_mus = TRUE,
  whole_region = TRUE)

create_polynomial_object(M = 1)
}
\arguments{
\item{M}{The number of the basis functions.}

\item{gamma}{Inverse width of radial basis function.}

\item{mus}{Optional centers of the RBF.}

\item{eq_spaced_mus}{Logical, if TRUE, equally spaced centers are created,
otherwise centers are created using \code{\link[stats]{kmeans}} algorithm.}

\item{whole_region}{Logical, indicating if the centers will be evaluated
equally spaced on the whole region, or between the min and max of the
observation values.}
}
\value{
A basis object of class 'rbf' or 'polynomial'.
}
\description{
These functions create different basis objects. These
objects can be used as input to complex functions in order to perform
computations depending on the class of the basis function.
}
\examples{
(obj <- create_rbf_object(M = 2))

#---------------------------------

(obj <- create_polynomial_object(M = 2))

}
\author{
C.A.Kapourani \email{C.A.Kapourani@ed.ac.uk}
}
\seealso{
\code{\link{eval_functions}}, \code{\link{bpr_optimize}}
}

