% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{create_genomic_region}
\alias{create_genomic_region}
\title{Create genomic regions from annotation data.}
\usage{
create_genomic_region(anno, chrom_size = NULL, is_centre = FALSE,
  upstream = -5000, downstream = 5000)
}
\arguments{
\item{anno}{A \code{\link[GenomicRanges]{GRanges}} object containing the
annotation data, this normally would be the output from
\code{\link{read_anno}} function.}

\item{chrom_size}{Object containing genome chromosome sizes, normally would
be the output of \code{\link{read_chrom_size}} function.}

\item{is_centre}{Logical, whether 'start' and 'end' locations are
pre-centred. If TRUE, the mean of the locations will be chosen as centre.
If FALSE, the 'start' will be chosen as the center; e.g. for genes the
'start' denotes the TSS and we use this as centre to obtain K-bp upstream
and downstream of TSS.}

\item{upstream}{Integer defining the length of bp upstream of 'centre' for
creating the genomic region.}

\item{downstream}{Integer defining the length of bp downstream of 'centre'
for creating the genomic region.}
}
\value{
A \code{\link[GenomicRanges]{GRanges}} object containing the genomic
  regions.

  The GRanges object contains two or three additional metadata column:
  \itemize{\item{\code{id}: Genomic region id.} \item{\code{centre}: Central
  location of each genomic region.} \item{\code{name}: (Optional) Genomic
  region name.} } This column can be accessed as follows:
  \code{granges_object$tss}
}
\description{
\code{create_genomic_region} creates genomic region from
  annotation data, using the central point of the annotation features as
  ground truth labels we create genomic regions \code{N} bp upstream and
  \code{M} bp downstream of central location.
}
\examples{
\dontrun{
# Download the files and change the working directory to that location
anno_file <- "name_of_anno_file"
annot_data <- read_anno(anno_file)
gen_region <- create_genomic_region(annot_data)

# Extract ID
id <- gen_region$id
}

}
\seealso{
\code{\link{create_met_region}}, \code{\link{read_anno}}
}
\author{
C.A.Kapourani \email{C.A.Kapourani@ed.ac.uk}
}
