% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_methyl_region.R
\name{create_methyl_region}
\alias{create_methyl_region}
\title{Create methylation regions for each gene promoter.}
\usage{
create_methyl_region(bs_data, prom_region, cpg_density = 1, sd_thresh = 0,
  ignore_strand = FALSE, fmin = -1, fmax = 1)
}
\arguments{
\item{bs_data}{\code{\link[GenomicRanges]{GRanges}} object containing the
BS-Seq data.}

\item{prom_region}{\code{\link[GenomicRanges]{GRanges}} object containing
promoter regions, i.e. N bp upstream and M bp downstream of TSS location.}

\item{cpg_density}{Optional integer defining the minimum number of CpGs that
have to be in a methylated region. Regions with less than \code{n} CpGs
are discarded.}

\item{sd_thresh}{Optional numeric defining the minimum standard deviation
of the methylation change in a region.}

\item{ignore_strand}{Logical, whether or not to ignore strand information.}

\item{fmin}{Optional minimum range value for region location scaling.}

\item{fmax}{Optional maximum range value for region location scaling.}
}
\value{
A \code{meth_regions} object containing the following information:
 \itemize{
   \item{ \code{meth_data}: A list containing methylation data, where each
     each entry in the list consists of an L X 3 dimensional matrix, where:
     \enumerate{
       \item{ 1st column: Contains the locations of CpGs relative to TSS,
         where the range (min, max) of possible values is given, by the
         inputs fmin and fmax.
       }
       \item{ 2nd column: Contains the total reads of the CpG in the
         corresponding location.}
       \item{ 3rd column: Contains the methylated reads of the CpG in the
         corresponding location.}
     }
   }
   \item{ \code{prom_ind}: A vector storing the corresponding promoter
     indices.}
 }
}
\description{
\code{create_methyl_region} creates methylation regions using BS-Seq and
 pre-processed annotated gene promoter regions. BS-Seq data give information
 for the methylated CpGs individually, and promoter data are used to locate
 the TSS of each gene and its promoter region.
}
\seealso{
\code{\link{read_rna_encode_caltech}},
 \code{\link{read_bs_encode_haib}},
 \code{\link{create_prom_region}}
}

