% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis_functions.R
\name{create_polynomial_object}
\alias{create_polynomial_object}
\title{Create a polynomial object}
\usage{
create_polynomial_object(M = 1)
}
\arguments{
\item{M}{The degree of the polynomial object.}
}
\value{
An object of class 'polynomial'
}
\description{
\code{create_polynomial_object} creates objects of class "polynomial". These
objects can be used as input to complex functions in order to perform
computations depending on the class of  basis function.
}
\examples{
(obj <- create_polynomial_object(M = 2))

}
\author{
C.A.Kapourani \email{C.A.Kapourani@ed.ac.uk}
}
\seealso{
\code{\link{create_rbf_object}}
}

