% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_prom_region.R
\name{create_prom_region}
\alias{create_prom_region}
\title{Create promoter regions from gene annotation data.}
\usage{
create_prom_region(annot_data, chrom_size = NULL, upstream = -100,
  downstream = 100)
}
\arguments{
\item{annot_data}{A GRanges object containing the gene annotation data.}

\item{chrom_size}{Optional data.frame containing the chromosome sizes.}

\item{upstream}{Integer defining the length of bp upstream of TSS.}

\item{downstream}{Integer defining the length of bp downstream of TSS.}
}
\value{
The created promoter regions stored in a
 \code{\link[GenomicRanges]{GRanges}} object.
}
\description{
\code{create_prom_region} creates promoter region from gene annotation data.
 Using the TSS of gene annotation data as ground truth labels we create
 promoter regions \code{N} bp upstream and \code{M} bp downstream of TSS.
}
\seealso{
\code{\link{create_methyl_region}}
}

