% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis_functions.R
\name{create_rbf_object}
\alias{create_rbf_object}
\title{Create an RBF object}
\usage{
create_rbf_object(M = 2, gamma = NULL, mus = NULL, eq_spaced_mus = TRUE,
  whole_region = TRUE)
}
\arguments{
\item{M}{The number of the RBFs.}

\item{gamma}{Inverse width of radial basis function.}

\item{mus}{Optional centers of the RBF.}

\item{eq_spaced_mus}{Logical, if TRUE, equally spaced centers are created,
otherwise centers are created using \code{\link[stats]{kmeans}} algorithm.}

\item{whole_region}{Logical, indicating if the centers will be evaluated
equally spaced on the whole region, or between the min and max of the
observation values.}
}
\value{
An object of class 'rbf'.
}
\description{
\code{create_rbf_object} creates objects of class "rbf". These objects can be
used as input to complex functions in order to perform computations depending
on the basis function.
}
\examples{
(obj <- create_rbf_object(M = 2))

}
\author{
C.A.Kapourani \email{C.A.Kapourani@ed.ac.uk}
}
\seealso{
\code{\link{create_polynomial_object}}
}

