% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_matrix.R
\name{design_matrix.polynomial}
\alias{design_matrix.polynomial}
\title{Create polynomial design matrices}
\usage{
\method{design_matrix}{polynomial}(x, obs, ...)
}
\arguments{
\item{x}{A basis object.}

\item{obs}{A vector of observations.}

\item{...}{Additional parameters}
}
\value{
A list containing the design matrix H and the basis object. The
 dimensions of the matrix H are N x (M+1), where N is the length of the
 observations, and M is the degree of the polynomial.
}
\description{
\code{design_matrix.polynomial} creates a design matrix H using polynomial
basis functions of degree M.
}
\examples{
obj <- polynomial.object(M=2)
obs <- c(0,.2,.5)
des_mat <- design_matrix(obj, obs)

}
\seealso{
\code{\link{polynomial_basis}}, \code{\link{polynomial.object}}
}

