% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{discard_bs_noise_reads}
\alias{discard_bs_noise_reads}
\title{Discard BS-Seq noisy reads}
\usage{
discard_bs_noise_reads(bs_data, min_bs_cov = 2, max_bs_cov = 1000)
}
\arguments{
\item{bs_data}{A Granges object containing the BS-Seq data.}

\item{min_bs_cov}{The minimum number of reads mapping to each CpG site.}

\item{max_bs_cov}{The maximum number of reads mapping to each CpG site.}
}
\value{
The reduced Granges object without noisy observations
}
\description{
\code{discard_bs_noise_reads} discards low coverage and (really) high reads
 from BS-Seq experiments. These reads can be thought as noise of the
 experiment.
}

