% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{discard_chr}
\alias{discard_chr}
\title{Discard selected chromosomes}
\usage{
discard_chr(x, chr_discarded = NULL)
}
\arguments{
\item{x}{The HTS data stored in a data.table object}

\item{chr_discarded}{A vector with chromosome names to be discarded.}
}
\value{
The reduced HTS data.
}
\description{
\code{discard_chr} Discards selected chromosomes
}

