% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{encode_expr}
\alias{encode_expr}
\title{Processed ENCODE expression data}
\format{Expression data in format as the output of \code{\link{read_expr}}
  fucntion, i.e. a \code{data.table} with two columns: "id", "expr".}
\usage{
encode_expr
}
\value{
Encode expression data
}
\description{
Small subset of ENCODE expression data already in pre-processed
  format, which are used as a case study for the vignette.
}
\seealso{
\code{\link{create_region_object}}, \code{\link{read_met}},
  \code{\link{read_anno}}, \code{\link{read_expr}}
}
\keyword{datasets}
