% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{encode_met}
\alias{encode_met}
\title{Processed ENCODE methylation data}
\format{A list object containing methylation regions and annotation data.
  This in general would be the output of the
  \code{\link{create_region_object}} function. It has the following two
  objects: \itemize{ \item{ \code{met}: A list containing the methylation
  regions, each element of the list is a different genomic region.} \item{
  \code{anno}: Corresponding annotation data.}}}
\usage{
encode_met
}
\value{
Encode methylation data
}
\description{
Small subset of ENCODE methylation data already in pre-processed
  format, which are used as a case study for the vignette.
}
\seealso{
\code{\link{create_region_object}}, \code{\link{read_met}},
  \code{\link{read_anno}}, \code{\link{read_expr}}
}
\keyword{datasets}
