% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_functions.R
\name{eval_function.polynomial}
\alias{eval_function.polynomial}
\title{Evaluate polynomial function}
\usage{
\method{eval_function}{polynomial}(x, obs, w, ...)
}
\arguments{
\item{x}{The basis function object.}

\item{obs}{Input / observation data.}

\item{w}{Vector of length M, containing the coefficients of an Mth-order
basis function.}

\item{...}{Optional additional parameters}
}
\value{
The polynomial function values.
}
\description{
Method for evaluating the polynomial function of degree M for observation
data obs and coefficients w.
}
\examples{
x <- polynomial.object(M=2)
obs <- c(1,2,3)
w <- c(0.1, 0.3, -0.6)
out <- eval_function(x, obs, w)

}

