% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_functions.R
\name{eval_function.rbf}
\alias{eval_function.rbf}
\title{Evaluate rbf function}
\usage{
\method{eval_function}{rbf}(x, obs, w, ...)
}
\arguments{
\item{x}{The basis function object.}

\item{obs}{Input / observation data.}

\item{w}{Vector of length M, containing the coefficients of an Mth-order
basis function.}

\item{...}{Optional additional parameters}
}
\value{
The rbf function values.
}
\description{
Method for evaluating the rbf function with M basis for observation data
obs and coefficients w.
}
\examples{
x <- create_rbf_object(M=2, mus = c(2,2.5))
obs <- c(1,2,3)
w <- c(0.1, 0.3, -0.6)
out <- eval_function(x, obs, w)

}
\author{
C.A.Kapourani \email{C.A.Kapourani@ed.ac.uk}
}

