% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_functions.R
\name{eval_function}
\alias{eval_function}
\title{Generic function for evaluating basis functions}
\usage{
eval_function(x, ...)
}
\arguments{
\item{x}{The basis function object.}

\item{...}{Additional parameters that will be passed to more specific
functions.}
}
\description{
Method for evaluating a basis function object of class x.
}
\examples{
x <- rbf.object(M=2, mus = c(2,2.5))
obs <- c(1,2,3)
w <- c(0.1, 0.3, -0.6)
out <- eval_function(x, obs, w)

#----------------

x <- polynomial.object(M=2)
obs <- c(1,2,3)
w <- c(0.1, 0.3, -0.6)
out <- eval_function(x, obs, w)

}
\seealso{
\code{\link{eval_probit_function}}, \code{\link{eval_function.rbf}},
 \code{\link{eval_function.polynomial}}
}

