% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_functions.R
\name{eval_probit_function}
\alias{eval_probit_function}
\title{Generic function for evaluating probit basis functions}
\usage{
eval_probit_function(x, ...)
}
\arguments{
\item{x}{The basis function object.}

\item{...}{Additional parameters that will be passed to more specific
functions.}
}
\value{
The probit transformed basis function values.
}
\description{
Method for evaluating the probit transformation of a basis function.
}
\examples{
x <- polynomial.object(M=2)
obs <- c(1,2,3)
w <- c(0.1, 0.3, -0.6)
out <- eval_probit_function(x, obs, w)

}
\seealso{
\code{\link{eval_function}}, \code{\link{eval_function.rbf}},
 \code{\link{eval_function.polynomial}}
}

