% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_fpkm}
\alias{extract_fpkm}
\title{Extract FPKM from string}
\usage{
extract_fpkm(x)
}
\arguments{
\item{x}{a string containing FPKM information}
}
\value{
The FPKM numeric value
}
\description{
\code{extract_fpkm} Extracts FPKM value from a string
}
\examples{
data <- 'gene_id "72"; transcr "ENST00000456328"; FPKM "0.0736851531";'
scaled <- extract_fpkm(data)

}

