% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_gene_name}
\alias{extract_gene_name}
\title{Extract gene name from string}
\usage{
extract_gene_name(x)
}
\arguments{
\item{x}{a string containing gene name information}
}
\value{
The gene name as a string
}
\description{
\code{extract_gene_name} Extracts gene name from a string
}
\examples{
data <- 'gene_name "Bnt1.1"; transcr "ENST00000456328"; FPKM "0.0736831";'
scaled <- extract_gene_name(data)

}

