% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gaussian_data}
\alias{gaussian_data}
\title{Synthetic Gaussian data}
\format{A list with 300 elements, where each element element is an L x 2
  matrix of observations, where: \describe{ \item{1st column}{locations of
  obseravtions} \item{2nd column}{methylation level} }}
\usage{
gaussian_data
}
\value{
Synthetic Bernoulli methylation data.
}
\description{
A synthetic dataset containinig 300 entries from Gaussian (i.e.
  linear) regression observations (e.g. M-values from array methylation
  data).
}
\seealso{
\code{\link{binomial_data}}, \code{\link{beta_data}},
  \code{\link{bernoulli_data}}
}
\keyword{datasets}
