% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gex_data}
\alias{gex_data}
\title{Synthetic expression data}
\format{A vector of length 600}
\usage{
gex_data
}
\value{
Synthetic gene expression data
}
\description{
Corresponding gene expression data for the
  \code{\link{meth_data}}
}
\seealso{
\code{\link{meth_data}}, \code{\link{bernoulli_data}},
  \code{\link{binomial_data}}, \code{\link{beta_data}},
  \code{\link{gaussian_data}}
}
\keyword{datasets}
