% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersect_cgi_tss.R
\name{intersect_cgi_tss}
\alias{intersect_cgi_tss}
\title{Keep CGIs that intersect with promoter regions}
\usage{
intersect_cgi_tss(cgi_file, rna_files, chrom_size_file = NULL,
  chr_discarded = NULL, upstream = -100, downstream = 100,
  unique_tss = FALSE)
}
\arguments{
\item{cgi_file}{The name of the CGI '.bed' formatted file to read
data values from.}

\item{rna_files}{The name of the RNA-Seq '.bed' formatted file to
read data values from.}

\item{chrom_size_file}{Optional name of the file containing genome
chromosome sizes}

\item{chr_discarded}{A vector with chromosome names to be discarded.}

\item{upstream}{Integer defining the length of bp upstream of TSS.}

\item{downstream}{Integer defining the length of bp downstream of TSS.}

\item{unique_tss}{Logical, indicating if CGIs should match only to unique
TSS, default false, so as to consider strand direction.}
}
\value{
The remaining CGIs which intersect with promoter regions stored in a
 \code{\link[GenomicRanges]{GRanges}} object.
}
\description{
\code{intersect_cgi_tss} keeps only CpG islands that overlap with promoter
 regions \code{N} bp upstream and \code{M} bp downstream of TSS.
}
\seealso{
\code{\link{create_prom_region}}
}

