% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{log_sum_exp}
\alias{log_sum_exp}
\title{Compute stable Log-Sum-Exp}
\usage{
log_sum_exp(x)
}
\arguments{
\item{x}{A vector of observations}
}
\value{
The logs-sum-exp value
}
\description{
\code{log_sum_exp} computes the log sum exp trick for avoiding numeric
underflow and have numeric stability in computations of small numbers.
}
\examples{
x <- c(0.001, 0.5, 2, 1.4, 1.5)
out <- log_sum_exp(x)

}
\references{
\url{https://hips.seas.harvard.edu/blog/2013/01/09/computing-log-sum-exp/}
}

