% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{minmax_scaling}
\alias{minmax_scaling}
\title{Compute the min-max scaling}
\usage{
minmax_scaling(data, xmin = NULL, xmax = NULL, fmin = -1, fmax = 1)
}
\arguments{
\item{data}{Vector with numeric data to be scaled.}

\item{xmin}{Optional minimum value, otherwise \code{min(data)} will be used.}

\item{xmax}{Optional maximum value, otherwise \code{max(data)} will be used.}

\item{fmin}{Optional minimum range value, default is -1.}

\item{fmax}{Optional maximum range value, default is 1.}
}
\value{
The scaled data in the given range, default is between (-1, 1). If
 xmin = xmax the input vector \code{data} is returned.
}
\description{
\code{minmax_scaling} normalizes a given vector using the the min-max
scaling method. More formally:
\deqn{scaled = \frac{data -x_{min}}{x_{max} - x_{min}} \times (f_{max} - f_{min}) + f_{min}}
}
\examples{
data <- c(-20, 0, 15, 20)
scaled <- minmax_scaling(data)

}

