% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpgex_cluster.R
\name{mpgex_cluster}
\alias{mpgex_cluster}
\title{Cluster similar methylation profiles}
\usage{
mpgex_cluster(x, K = 3, pi_k = NULL, w = NULL, basis = NULL,
  em_max_iter = 100, epsilon_conv = 1e-04, opt_method = "CG",
  opt_itnmax = 100, init_opt_itnmax = 100, is_parallel = TRUE,
  no_cores = NULL, is_verbose = FALSE)
}
\arguments{
\item{x}{A list of elements of length N, where each element is an L x 3
matrix of observations, where 1st column contains the locations. The 2nd
and 3rd columns contain the total trials and number of successes at the
corresponding locations, repsectively.}

\item{K}{Integer denoting the number of clusters K.}

\item{pi_k}{Vector of length K, denoting the mixing proportions.}

\item{w}{A MxK matrix, where each column contains the basis function
coefficients for the corresponding cluster.}

\item{basis}{A 'basis' object. E.g. see \code{\link{create_rbf_object}}.}

\item{em_max_iter}{Integer denoting the maximum number of EM iterations.}

\item{epsilon_conv}{Numeric denoting the convergence parameter for EM.}

\item{opt_method}{The optimization method to be used. See
\code{\link[stats]{optim}} for possible methods. Default is 'CG'.}

\item{opt_itnmax}{Optional argument giving the maximum number of iterations
for the corresponding method. See \code{\link[stats]{optim}} for details.}

\item{init_opt_itnmax}{Optimization iterations for obtaining the initial EM
parameter values.}

\item{is_parallel}{Logical, indicating if code should be run in parallel.}

\item{no_cores}{Number of cores to be used, default is max_no_cores - 2.}

\item{is_verbose}{Logical, print results during EM iterations}
}
\description{
\code{mpgex_cluster} is a wrapper function that clusters similar methylation
profiles using the EM algorithm. Initially, it performs parameter checking,
and initializes main parameters, such as mixing proportions, basis function
coefficients, then the EM algorithm is applied and finally model selection
metrics are calculated, such as BIC and AIC.
}
\examples{
ex_data <- meth_data
mpgex_clust <- mpgex_cluster(x = ex_data, em_max_iter = 3, is_parallel = FALSE,
                             opt_itnmax = 10)

}
\author{
C.A.Kapourani \email{C.A.Kapourani@ed.ac.uk}
}

