% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpgex_regr.R
\name{mpgex_regr}
\alias{mpgex_regr}
\title{Predict gene expression from methylation profiles}
\usage{
mpgex_regr(formula = NULL, x, y, model_name = "svm", w = NULL,
  basis = NULL, train_ind = NULL, train_perc = 0.7, fit_feature = NULL,
  cpg_dens_feat = FALSE, opt_method = "CG", opt_itnmax = 500,
  is_parallel = TRUE, no_cores = NULL, is_summary = TRUE)
}
\arguments{
\item{formula}{An object of class \code{\link[stats]{formula}} needed when
calling the \code{\link[stats]{lm}} function for performing linear
regression. If NULL, the simple linear regression method is used.}

\item{x}{The binomial distributed observations, which has to be a list where
each element is an L x 3 dimensional matrix.}

\item{y}{Corresponding gene expression data for each element of the list x}

\item{model_name}{A charcter denoting the regression model.}

\item{w}{Optional vector of initial parameter / coefficient values.}

\item{basis}{Optional basis function object, default is
\code{\link{create_rbf_object}}.}

\item{train_ind}{Optional vector containing the indices for the train set.}

\item{train_perc}{Optional parameter for defining the percentage of the
dataset to be used for training set, the remaining will be the test set.}

\item{fit_feature}{Additional feature on how well the profile fits the
methylation data.}

\item{cpg_dens_feat}{Additional feature for the CpG density across the
promoter region.}

\item{opt_method}{Parameter for defining the method to be used in the
optimization procedure, see \code{\link[stats]{optim}}.}

\item{opt_itnmax}{Optional parameter for defining the max number of
iterations of the optimization procedure, see \code{\link[stats]{optim}}.}

\item{is_parallel}{Logical, indicating if code should be run in parallel.}

\item{no_cores}{Number of cores to be used, default is max_no_cores - 1.}

\item{is_summary}{Logical, print the summary statistics.}
}
\value{
An mpgex object consisting of the following elements:
}
\description{
\code{mpgex_regr} is a function that wraps all the necessary subroutines for
performing predictions on gene expressions. Initially, it optimizes the
parameters of the basis functions so as to learn the methylation profiles.
Then uses the learned parameters / coefficients of the basis functions as
input features for performing linear regression in order to predict/regress
the corresponding gene expression data.
}
\examples{
obs <- meth_data
y   <- gex_data
basis <- create_rbf_object(M = 5)
out   <- mpgex_regr(x = obs, y = y, basis = basis, is_parallel = FALSE,
                    opt_itnmax = 10)

}
\author{
C.A.Kapourani \email{C.A.Kapourani@ed.ac.uk}
}
\seealso{
\code{\link{bpr_optim}}, \code{\link{create_polynomial_object}},
  \code{\link{create_rbf_object}}
}

