% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_functions.R
\name{old_plot_fitted_profiles}
\alias{old_plot_fitted_profiles}
\title{(DEPRECATED) Plot the fit of methylation profiles across a region}
\usage{
old_plot_fitted_profiles(
  region,
  X,
  fit_prof,
  fit_mean = NULL,
  title = "Gene promoter",
  ...
)
}
\arguments{
\item{region}{Promoter region number}

\item{X}{Methylation data observations}

\item{fit_prof}{Fitted profile}

\item{fit_mean}{Fitted mean function}

\item{title}{Title of the plot}

\item{...}{Additional parameters}
}
\value{
The figure to be plotted in the device.
}
\description{
(DEPRECATED) \code{plot_fitted_profiles} is a simple function
  for plotting the methylation data across a give region, together with the
  fit of the methylation profiles.
}
\examples{
# Fit methylation profiles using 3 RBFs
obs <- meth_data
y   <- gex_data
basis <- create_rbf_object(M = 3)
out   <- bpr_predict_wrap(x = obs, y = y, basis = basis,
                          is_parallel = FALSE, opt_itnmax = 5)

# Create the plot
old_plot_fitted_profiles(region = 16, X = meth_data, fit_prof = out)

}
\seealso{
\code{\link{old_plot_cluster_prof}}
}
\author{
C.A.Kapourani \email{C.A.Kapourani@ed.ac.uk}
}
