% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_fitted_profiles}
\alias{plot_fitted_profiles}
\title{Plot the fit of methylation profiles across a region}
\usage{
plot_fitted_profiles(region, X, fit_prof, fit_mean = NULL,
  title = "Gene promoter", ...)
}
\arguments{
\item{region}{Promoter region number}

\item{X}{Methylation data observations}

\item{fit_prof}{Fitted profile}

\item{fit_mean}{Fitted mean function}

\item{title}{Title of the plot}

\item{...}{Additional parameters}
}
\value{
The figure to be plotted in the device.
}
\description{
\code{plot_fitted_profiles} is a simple function for plotting the methylation
data across a give region, together with the fit of the methylation profiles.
}
\examples{
# Fit methylation profiles using 8 RBFs
obs <- meth_data
y   <- gex_data
basis <- create_rbf_object(M = 8)
out   <- bpr_predict_wrap(x = obs, y = y, basis = basis,
                          is_parallel = FALSE, opt_itnmax = 10)

# Create the plot
plot_fitted_profiles(region = 16, X = meth_data, fit_prof = out)

}
\author{
C.A.Kapourani \email{C.A.Kapourani@ed.ac.uk}
}
\seealso{
\code{\link{plot_cluster_prof}}, \code{\link{plot_scatter_gex}},
  \code{\link{boxplot_cluster_gex}}
}

