% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_scatter_gex}
\alias{plot_scatter_gex}
\title{Scatter plot of predicted vs measured gene expression levels}
\usage{
plot_scatter_gex(bpr_predict_obj, main_lab = "Methylation Profile",
  is_margins = TRUE)
}
\arguments{
\item{bpr_predict_obj}{The output of the \code{bpr_predict_wrap} function.}

\item{main_lab}{The title of the plot}

\item{is_margins}{Use specific margins or not.}
}
\value{
The figure to be plotted in the device.
}
\description{
\code{plot_scatter_gex} creates a scatter plot of predicted gene expression
values on the x-axis versus the measured gene expression values on the
y-axis.
}
\examples{
# Fit methylation profiles using 8 RBFs
obs <- meth_data
y   <- gex_data
basis <- create_rbf_object(M = 8)
res   <- bpr_predict_wrap(x = obs, y = y, basis = basis,
                          is_parallel = FALSE, opt_itnmax = 10)

# Create the scatter plot
plot_scatter_gex(bpr_predict_obj = res)

}
\author{
C.A.Kapourani \email{C.A.Kapourani@ed.ac.uk}
}
\seealso{
\code{\link{plot_cluster_prof}}, \code{\link{plot_fitted_profiles}},
  \code{\link{boxplot_cluster_gex}}
}

