% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis_functions.R
\name{polynomial_basis}
\alias{polynomial_basis}
\title{Apply polynomial basis function.}
\usage{
polynomial_basis(X, M = 1)
}
\arguments{
\item{X}{The input data, either a scalar, vector or matrix.}

\item{M}{Integer, denoting the degree of the polynomial basis that will be
applied to the input X. M should not be negative.}
}
\value{
Input X, after being transformed from the polynomial basis function.
}
\description{
Applies the polynomial basis function of degree M to the input X.
}
\examples{
out <- polynomial_basis(X = 3, M = 2)
#
out <- polynomial_basis(X = c(2, 3, 4), M = 2)

}
\seealso{
\code{\link{polynomial.object}}, \code{\link{rbf_basis}}
}

