% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool_bs_encode_haib_rep.R
\name{pool_bs_encode_haib_rep}
\alias{pool_bs_encode_haib_rep}
\title{Read and pool replicates from Bismark Cov formatted BS-Seq data}
\usage{
pool_bs_encode_haib_rep(files, chr_discarded = NULL)
}
\arguments{
\item{files}{A vector of filenames containing replicate experiments.}

\item{chr_discarded}{A vector with chromosome names to be discarded.}
}
\value{
A \code{\link[GenomicRanges]{GRanges}} object.
}
\description{
\code{pool_bs_bismark_cov_rep} reads and pools replicate methylation data
 from  BS-Seq experiments that are in Bismark cov format.
}

