% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blm.R
\name{predict.blm}
\alias{predict.blm}
\title{Make predictions using the Basis Linear Model}
\usage{
\method{predict}{blm}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{Object of class \code{\link{blm}}.}

\item{newdata}{An optional data frame in which to look for variables with
which to predict. If omitted, the fitted values are used.}

\item{...}{Optional additional parameters.}
}
\value{
\code{predict.blm} produces a vector of predictions.
}
\description{
S3 method for class blm, that makes predictions for \code{newdata} using
the Basis Linear Model. This funciton is similar to \code{\link[stats]{lm}},
however currently it does not provide the \code{\link[stats]{formula}}
functionality.
}

