% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_model_gex.R
\name{predict_model_gex}
\alias{predict_model_gex}
\title{Predict gene expression model from methylation profiles}
\usage{
predict_model_gex(model, test, is_summary = TRUE)
}
\arguments{
\item{model}{The fitted regression model, i.e. the output of
\code{\link{train_model_gex}}.}

\item{test}{The testing data.}

\item{is_summary}{Logical, print the summary statistics.}
}
\value{
A list containing the following elements: \itemize{ \item
  \code{test_pred}: The predicted values for the test data. \item
  \code{test_errors}: The test error metrics. }
}
\description{
\code{predict_model_gex} makes predictions of gene expression levels using a
model trained on higher order methylation features extracted from specific
genomic regions.
}
\examples{
# Create synthetic data
train_data <- data.frame(x = rnorm(20), y=rnorm(20, 1, 4))
test_data <- data.frame(x = rnorm(20), y=rnorm(20, 1, 3))

# Train the model
train_model <- train_model_gex(formula = y~., train = train_data)

# Make predictions
res <- predict_model_gex(model = train_model$gex_model, test = test_data)

}
\author{
C.A.Kapourani \email{C.A.Kapourani@ed.ac.uk}
}
\seealso{
\code{\link{train_model_gex}}
}

