% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_bs_bismark_cov.R
\name{preprocess_bs_bismark_cov}
\alias{preprocess_bs_bismark_cov}
\title{Process BS-Seq data in Bismark cov format}
\usage{
preprocess_bs_bismark_cov(files, chr_discarded = NULL, min_bs_cov = 2,
  max_bs_cov = 1000)
}
\arguments{
\item{files}{A vector of filenames containing replicate experiments.}

\item{chr_discarded}{A vector with chromosome names to be discarded.}

\item{min_bs_cov}{Minimum number of reads mapping to each CpG site.}

\item{max_bs_cov}{Maximum number of reads mapping to each CpG site.}
}
\value{
A \code{\link[GenomicRanges]{GRanges}} object.
}
\description{
\code{preprocess_bs_bismark_cov} pre-processes BS-Seq files in Bismark Cov
 format. If a vector of files is given, the replicates are pooled together,
 then the total reads are extracted and finally noisy reads are discarded.
}

