% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blm.R
\name{print.blm}
\alias{print.blm}
\title{Print the output of the Basis Linear Model}
\usage{
\method{print}{blm}(x, ...)
}
\arguments{
\item{x}{Object of class \code{\link{blm}}.}

\item{...}{Optional additional parameters.}
}
\description{
S3 method for class blm, that prints the output of the Basis Linear Model
in a similar way to \code{\link[stats]{lm}}.
}

