% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blm.R
\name{print.summary.blm}
\alias{print.summary.blm}
\title{Print summary output of the Basis Linear Model}
\usage{
\method{print}{summary.blm}(x, ...)
}
\arguments{
\item{x}{Object of class \code{\link{summary.blm}}.}

\item{...}{Optional additional parameters.}
}
\description{
S3 method for class summary.blm, that prints the summary of the Basis Linear
Model in a similar way to \code{\link[stats]{lm}}.
}

