% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_haib_caltech.R
\name{process_haib_caltech}
\alias{process_haib_caltech}
\title{Wrapper method for processing ENCODE HAIB and Caltech HTS data}
\usage{
process_haib_caltech(bs_files, rna_files, chrom_size_file = NULL,
  chr_discarded = NULL, upstream = -100, downstream = 100,
  min_bs_cov = 0, max_bs_cov = 1000, cpg_density = 1, sd_thresh = 0,
  ignore_strand = FALSE, fmin = -1, fmax = 1)
}
\arguments{
\item{bs_files}{The name of the BS-Seq '.bed' formatted file to read
data values from.}

\item{rna_files}{The name of the RNA-Seq '.bed' formatted file to
read data values from.}

\item{chrom_size_file}{Optional name of the file containing genome
chromosome sizes}

\item{chr_discarded}{A vector with chromosome names to be discarded.}

\item{upstream}{Integer defining the length of bp upstream of TSS.}

\item{downstream}{Integer defining the length of bp downstream of TSS.}

\item{min_bs_cov}{The minimum number of reads mapping to each CpG site.}

\item{max_bs_cov}{The maximum number of reads mapping to each CpG site.}

\item{cpg_density}{Optional integer defining the minimum number of CpGs that
have to be in a methylated region. Regions with less than \code{n} CpGs are
discarded.}

\item{sd_thresh}{Optional numeric defining the minimum standard deviation of
the methylation change in a region. This is used to filter regions with no
methylation change.}

\item{ignore_strand}{Logical, whether or not to ignore strand information.}

\item{fmin}{Optional minimum range value for region location scaling. Under
this version, this parameter should be left to its default value.}

\item{fmax}{Optional maximum range value for region location scaling. Under
this version, this parameter should be left to its default value.}
}
\value{
A \code{processHTS} object which contains among others the following
 information:
 \itemize{
   \item{ \code{methyl_region}: A list containing the methylation regions,
     where each each entry in the list consists of an L X 3 dimensional
     matrix, where:
     \enumerate{
       \item{ 1st col: Contains the locations of the CpGs relative to TSS,
         where the range (min, max) of possible values is given, by the
         inputs fmin and fmax.
       }
       \item{ 2nd col: The total reads of the CpG in the corresponding
         location.}
       \item{ 3rd col: The methylated reads of the CpG in the corresponding
         location.}
     }
   }
   \item{ \code{prom_region}: A \code{\link[GenomicRanges]{GRanges}} object
     containing corresponding annotated promoter regions for each entry of
     the \code{methyl_region} list..

   }
   \item{ \code{rna_data}: A \code{\link[GenomicRanges]{GRanges}} object
     containing the corresponding RNA-Seq data for each entry of the
     \code{methyl_region} list.}
 }
}
\description{
\code{process_haib_caltech} is a wrapper method for processing HTS data and
returning the methylation promoter regions and the corresponding gene
expression data for those promoter regions.
}

