% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis_functions.R
\name{rbf.object}
\alias{rbf.object}
\title{Create an RBF object}
\usage{
rbf.object(M = 2, gamma = NULL, mus = NULL, eq_spaced_mus = TRUE,
  whole_region = TRUE)
}
\arguments{
\item{M}{The number of the RBFs.}

\item{gamma}{Inverse width of radial basis function.}

\item{mus}{Optional centers of the RBF.}

\item{eq_spaced_mus}{Logical, if TRUE, equally spaced centers are created,
otherwise centers are created using \code{\link[stats]{kmeans}} algorithm.}

\item{whole_region}{Logical, indicating if the centers will be evaluated
equally spaced on the whole region, or between the min and max of the
observation values.}
}
\value{
An object of type 'RBF'.
}
\description{
\code{rbf.object} creates objects of class "rbf". These objects can be used
as input to complex functions in order to perform computations depending on
the basis function. E.g see \code{\link{design_matrix}} generic function.
}
\examples{
(obj <- rbf.object(M = 2))

}
\seealso{
\code{\link{design_matrix}}, \code{\link{polynomial.object}}
}

