% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis_functions.R
\name{rbf_basis}
\alias{rbf_basis}
\title{Apply the radial basis function}
\usage{
rbf_basis(X, mus, gamma = 1)
}
\arguments{
\item{X}{Input data.}

\item{mus}{Centers from where we should compute the distance of the data X.}

\item{gamma}{Inverse width of radial basis function.}
}
\value{
Input X, after being transformed from the RBF.
}
\description{
Apply the RBF function to the input X.
}
\examples{
out <- rbf_basis(X = c(1,2), mus = c(1,1))
#
out <- rbf_basis(X = c(1,2), mus = c(1,1), gamma = 0.1)

}
\seealso{
\code{\link{rbf.object}}, \code{\link{polynomial_basis}}
}

