% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{read_anno}
\alias{read_anno}
\title{Read annotation file}
\usage{
read_anno(file, chrom_size_file = NULL, chr_discarded = NULL,
  is_centre = FALSE, upstream = -5000, downstream = 5000)
}
\arguments{
\item{file}{File name.}

\item{chrom_size_file}{Optional file name to read genome chromosome sizes.}

\item{chr_discarded}{Optional vector with chromosomes to be discarded.}

\item{is_centre}{Logical, whether 'start' and 'end' locations are
pre-centred. If TRUE, the mean of the locations will be chosen as centre.
If FALSE, the 'start' will be chosen as the center; e.g. for genes the
'start' denotes the TSS and we use this as centre to obtain K-bp upstream
and downstream of TSS.}

\item{upstream}{Integer defining the length of bp upstream of 'centre' for
creating the genomic region.}

\item{downstream}{Integer defining the length of bp downstream of 'centre'
for creating the genomic region.}
}
\value{
A \code{\link[GenomicRanges]{GRanges}} object.

  The GRanges object contains two or three additional metadata columns:
  \itemize{ \item{ \code{id}: Feature ID, e.g. Ensembl IDs for each gene.}
  \item{ \code{centre}: The central (middle) location of the genomic region.
  This is required when transforming 'methylation regions' in the (-1, 1)
  interval, the 'centre' would be at 0.} \item{\code{name} (Optional) the
  feature name.} } These columns can be accessed as follows:
  \code{granges_obj$id}
}
\description{
\code{read_anno} reads a file containing annotation data, which
  will be used to select genomic regions of interest for downstream analysis.
  The annotation file format is the following: "chromosome", "start", "end",
  "strand", "id", "name" (optional). Read the Important section below for
  more details.
}
\section{Important}{
 \itemize{ \item{When having strand information, and we
  are in the antisense strand ("-"), the 'start' is automatically switched
  with the 'end' location.} \item{Columns should be in tab delimited format.}
  \item{ The "name" feature is optional.} \item{When there is no "strand"
  info, this can be set to "*".}}
}

\examples{
\dontrun{
# Download the files and change the working directory to that location
file <- "name_of_anno_file"
anno_dt <- read_anno(file)

# Extract feature id
anno_ids <- anno_dt$id
}

}
\seealso{
\code{\link{read_met}}
}
\author{
C.A.Kapourani \email{C.A.Kapourani@ed.ac.uk}
}
