% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_encode_cgi.R
\name{read_encode_cgi}
\alias{read_encode_cgi}
\title{Read file containing CpG island locations}
\usage{
read_encode_cgi(file, is_GRanges = TRUE)
}
\arguments{
\item{file}{The name of the file to read data values from.}

\item{is_GRanges}{Logical: if TRUE a GRanges object is returned, otherwise a
data.frame object is returned.}
}
\value{
A \code{\link[GenomicRanges]{GRanges}} object if \code{is_GRanges} is
  TRUE, otherwise a \code{\link[data.table]{data.table}} object.

  The GRanges object contains one additional metadata column: \itemize{ \item
  \code{cgi_id}: Unique ID of the CpG Island. } This column can be accessed
  as follows: \code{granges_object$cgi_id}
}
\description{
\code{read_encode_cgi} reads a file containing CpG island (CGI) locations in
the human genome using the \code{\link[data.table]{fread}} function.
}
\examples{
\dontrun{
# Download the file and change the working directory to that location
file <- "name_of_CGI_file")
cgi_data <- read_encode_cgi(file)

# Extract the CGI ID
cgi_id <- cgi_data$cgi_id
}

}
\author{
C.A.Kapourani \email{C.A.Kapourani@ed.ac.uk}
}
\seealso{
\code{\link{read_rna_encode_caltech}},
  \code{\link{read_bs_encode_haib}}
}

