% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rna_encode_caltech.R
\name{read_rna_encode_caltech}
\alias{read_rna_encode_caltech}
\title{Read file containing ENCODE Caltech \code{bed} formatted RNA-Seq data}
\usage{
read_rna_encode_caltech(file, chr_discarded = NULL, is_GRanges = TRUE)
}
\arguments{
\item{file}{The name of the file to read data values from.}

\item{chr_discarded}{A vector with chromosome names to be discarded.}

\item{is_GRanges}{Logical: if TRUE a \code{\link[GenomicRanges]{GRanges}}
object is returned, otherwise a \code{\link[data.table]{data.table}} object
is returned.}
}
\value{
a \code{\link[GenomicRanges]{GRanges}} object if \code{is_GRanges}
 is TRUE, otherwise a \code{\link[data.table]{data.table}} object.
}
\description{
\code{read_rna_encode_caltech} reads a file containing promoter annotation
data together with gene expression levels from RNA-Seq experiments using the
\code{\link{scan}} function. The RNA-Seq file should be in ENCODE Caltech
\code{bed} format, e.g. use \code{gtf2bed} tool if your initial file is in
\code{gtf} format.
}
\seealso{
\code{\link{read_chrom_size}}, \code{\link{read_bs_encode_haib}}
}

