% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpr.R
\name{sum_weighted_bpr_lik}
\alias{sum_weighted_bpr_lik}
\title{Sum of weighted BPR log likelihoods}
\usage{
sum_weighted_bpr_lik(w, x, des_mat, post_prob, is_NLL = TRUE)
}
\arguments{
\item{w}{A vector of parameters (i.e. coefficients of the basis functions)}

\item{x}{A list of elements of length N, where each element is an L x 3
matrix of observations, where 1st column contains the locations. The 2nd
and 3rd columns contain the total trials and number of successes at the
corresponding locations, repsectively.}

\item{des_mat}{A list of length N, where each element contains the
\code{L x M} design matrices, where L is the number of observations and M
the number of basis functions.}

\item{post_prob}{A vector of length N containing the posterior probabilities
fore each element of list x, respectively.}

\item{is_NLL}{Logical, indicating if the Negative Log Likelihood should be
returned.}
}
\value{
The weighted sum of BPR log likelihoods
}
\description{
\code{sum_weighted_bpr_lik} computes the sum of the BPR log likelihoods for
each elements of x, and then weights them by the corresponding posterior
probabilities. This function is mainly used for the M-Step of the EM
algorithm \code{\link{bpr_EM}}.
}
\seealso{
\code{\link{bpr_likelihood}}, \code{\link{bpr_EM}}
}

