% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blm.R
\name{summary.blm}
\alias{summary.blm}
\title{Summary output of the Basis Linear Model}
\usage{
\method{summary}{blm}(object, ...)
}
\arguments{
\item{object}{Object of class \code{\link{blm}}.}

\item{...}{Optional additional parameters.}
}
\value{
A summary.blm object
}
\description{
S3 method for class blm, that computes the summary of the Basis Linear
Model in a similar way to \code{\link[stats]{lm}}.
}

