% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_model_gex.R
\name{train_model_gex}
\alias{train_model_gex}
\title{Train model for gene expression data from methylation profiles}
\usage{
train_model_gex(formula = NULL, model_name = "svm", train,
  is_summary = TRUE)
}
\arguments{
\item{formula}{An object of class \code{\link[stats]{formula}} needed when
calling the the model, e.g. \code{\link[stats]{lm}} function for performing
linear regression. If NULL, the simple linear regression model is used.}

\item{model_name}{A charcter denoting the regression model}

\item{train}{The training data.}

\item{is_summary}{Logical, print the summary statistics.}
}
\value{
A list containing the following elements:
\itemize{
 \item \code{formula} the formula that was used.
 \item \code{gex_model} the fitted model.
 \item \code{train_pred} the training predicted values.
 \item \code{train_errors} the training error metrics.
}
}
\description{
\code{train_model_gex} trains a regression model for gene expression data
 from methylation profiles.
}
\seealso{
\code{\link{calculate_errors}}, \code{\link{predict_model_gex}}
}

