% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_matrix.R
\name{design_matrix}
\alias{design_matrix}
\title{Generic function for creating design matrices}
\usage{
design_matrix(x, ...)
}
\arguments{
\item{x}{A basis function object.}

\item{...}{Additional parameters.}
}
\description{
This is a generic function which calls the appropriate methods depending on
the class of the object \code{x}.
}
\examples{
obj <- create_rbf_object(M=2)
obs <- c(0,.2,.5)
polyn <- design_matrix(obj, obs)

#----------------

obj <- create_polynomial_object(M=2)
obs <- c(0,.2,.5)
rbf <- design_matrix(obj, obs)

}
\author{
C.A.Kapourani \email{C.A.Kapourani@ed.ac.uk}
}
\seealso{
\code{\link{create_polynomial_object}},
  \code{\link{create_rbf_object}}, \code{\link{design_matrix.polynomial}},
  \code{\link{design_matrix.rbf}}
}

