% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_functions.R
\name{eval_functions}
\alias{eval_function}
\alias{eval_function.polynomial}
\alias{eval_function.rbf}
\alias{eval_functions}
\alias{eval_probit_function}
\title{Evaluate basis functions}
\usage{
eval_probit_function(x, ...)

eval_function(x, ...)

\method{eval_function}{rbf}(x, obs, w, ...)

\method{eval_function}{polynomial}(x, obs, w, ...)
}
\arguments{
\item{x}{The basis function object.}

\item{...}{Optional additional parameters}

\item{obs}{Observation data.}

\item{w}{Vector of length M, containing the coefficients of an
M\eqn{^{th}}-order basis function.}
}
\value{
The evaluated function values.

  NOTE that the \code{eval_probit_function} computes the probit transformed
  basis function values.
}
\description{
Method for evaluating an M basis function model with observation
  data \code{obs} and coefficients \code{w}.
}
\examples{
# Evaluate the probit transformed basis function values
x <- create_rbf_object(M=2)
obs <- c(1,2,3)
w <- c(0.1, 0.3, -0.6)
out <- eval_probit_function(x, obs, w)

# -------------------------

# Evaluate the RBF basis function values
x <- create_rbf_object(M=2, mus = c(2,2.5))
obs <- c(1,2,3)
w <- c(0.1, 0.3, -0.6)
out <- eval_function(x, obs, w)

# -------------------------

# Evaluate the Polynomial basis function values
x <- create_polynomial_object(M=2)
obs <- c(1,2,3)
w <- c(0.1, 0.3, -0.6)
out <- eval_function(x, obs, w)

}
\author{
C.A.Kapourani \email{C.A.Kapourani@ed.ac.uk}
}
\seealso{
\code{\link{create_basis}}
}

