% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gex_data}
\alias{gex_data}
\title{Synthetic data for mpgex package}
\format{A vector of length 600}
\usage{
gex_data
}
\description{
Corresponding gene expression data for the 'meth_data'
}
\keyword{datasets}

