% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meth_data}
\alias{meth_data}
\title{Synthetic bulk methylation data}
\format{
A list with 600 elements, where each element element is an L x 3
  matrix of observations, where: \describe{ \item{1st column}{locations of
  obseravtions} \item{2nd column}{total trials} \item{3rd column}{number of
  successes} }
}
\usage{
meth_data
}
\value{
Synthetic bulk methylation data
}
\description{
A synthetic dataset containinig 600 entries.
}
\seealso{
\code{\link{gex_data}}, \code{\link{bernoulli_data}},
  \code{\link{binomial_data}}, \code{\link{beta_data}},
  \code{\link{gaussian_data}}
}
\keyword{datasets}
