% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_model_gex.R
\name{predict_model_gex}
\alias{predict_model_gex}
\title{Predict gene expression data from methylation profiles}
\usage{
predict_model_gex(model, test, is_summary = TRUE)
}
\arguments{
\item{model}{The regression model learned during training}

\item{test}{The test data.}

\item{is_summary}{Logical, print the summary statistics.}
}
\value{
A list containing the following elements:
\itemize{
 \item \code{test_pred} the test predicted values.
 \item \code{test_errors} the test error metrics.
}
}
\description{
\code{predict_model_gex} predicts gene expression data from methylation
 profiles, using a regression model. The same method can be used to predict
 differential expression from differential methylation profiles.
}
\seealso{
\code{\link{train_model_gex}}
}

