% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_bs_seq.R
\name{preprocess_bs_seq}
\alias{preprocess_bs_seq}
\title{Pre-process BS-Seq data in any given format}
\usage{
preprocess_bs_seq(files, file_format = "encode_rrbs", chr_discarded = NULL,
  min_bs_cov = 2, max_bs_cov = 1000)
}
\arguments{
\item{files}{A vector of filenames containing replicate experiments.}

\item{file_format}{A string denoting the file format that the BS-Seq data are
stored. Current version allows "\code{encode_rrbs}" or "\code{bismark_cov}"
formats.}

\item{chr_discarded}{A vector with chromosome names to be discarded.}

\item{min_bs_cov}{Minimum number of reads mapping to each CpG site.}

\item{max_bs_cov}{Maximum number of reads mapping to each CpG site.}
}
\value{
A \code{\link[GenomicRanges]{GRanges}} object. The GRanges object
  contains two additional metadata columns: \itemize{ \item
  \code{total_reads}: total reads mapped to each genomic location. \item
  \code{meth_reads}: methylated reads mapped to each genomic location. }
  These columns can be accessed as follows: \code{granges_object$total_reads}
}
\description{
\code{preprocess_bs_seq} is a general function for reading and preprocessing
BS-Seq data. If a vector of files is given, these are considered as
replicates and are pooled together. Finally, noisy reads are discarded.
}
\section{Additional Info}{
 Information about the file formats can be found in
  the following links:

  Encode RRBS format:
  \url{http://rohsdb.cmb.usc.edu/GBshape/cgi-bin/hgTables?db=hg19&hgta_group=regulation&hgta_track=wgEncodeHaibMethylRrbs&hgta_table=wgEncodeHaibMethylRrbsBcbreast0203015BiochainSitesRep2&hgta_doSchema=describe+table+schema}

  Bismark Cov format: \url{http://rnbeads.mpi-inf.mpg.de/data/RnBeads.pdf}
}
\examples{
\dontrun{
# Download the files and change the working directory to that location
files <- c("name_of_bs_encode_file1", "name_of_bs_encode_file2")
bs_data <- preprocess_bs_seq(files, file_format = "encode_rrbs")

# Extract the total reads and methylated reads
total_reads <- bs_data$total_reads
meth_reads <- bs_data$meth_reads
}

}
\author{
C.A.Kapourani \email{C.A.Kapourani@ed.ac.uk}
}
\seealso{
\code{\link{read_bs_bismark_cov}}, \code{\link{read_bs_encode_haib}}
  \code{\link{pool_bs_seq_rep}}
}

