% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bs_bismark_cov.R
\name{read_bs_bismark_cov}
\alias{read_bs_bismark_cov}
\title{Read file containing Bismark Cov formatted BS-Seq data}
\usage{
read_bs_bismark_cov(file, chr_discarded = NULL, is_GRanges = TRUE)
}
\arguments{
\item{file}{The name of the file to read data values from.}

\item{chr_discarded}{A vector with chromosome names to be discarded.}

\item{is_GRanges}{Logical: if TRUE a \code{\link[GenomicRanges]{GRanges}}
object is returned, otherwise a \code{\link[data.table]{data.table}} object
is returned.}
}
\value{
a \code{\link[GenomicRanges]{GRanges}} object if \code{is_GRanges}
 is TRUE, otherwise a \code{\link[data.table]{data.table}} object.
}
\description{
\code{read_bs_bismark_cov} reads a file containing methylation data from
 BS-Seq experiments using the \code{\link[data.table]{fread}} function.
 The BS-Seq file should be in Bismark Cov format.
}
\examples{
\donttest{
# Download the files and change the working directory to that location
file <- "name_of_bismark_file"
rrbs <- read_bs_bismark_cov(file)}

}
\author{
C.A.Kapourani \email{C.A.Kapourani@ed.ac.uk}
}
\references{
\url{http://rnbeads.mpi-inf.mpg.de/data/RnBeads.pdf}
}
\seealso{
\code{\link{pool_bs_bismark_cov_rep}},
 \code{\link{preprocess_bs_bismark_cov}}
}

