% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bs_encode_haib.R
\name{read_bs_encode_haib}
\alias{read_bs_encode_haib}
\title{Read file containing ENCODE HAIB \code{bed} formatted BS-Seq data}
\usage{
read_bs_encode_haib(file, chr_discarded = NULL, is_GRanges = TRUE)
}
\arguments{
\item{file}{The name of the file to read data values from.}

\item{chr_discarded}{A vector with chromosome names to be discarded.}

\item{is_GRanges}{Logical: if TRUE a \code{\link[GenomicRanges]{GRanges}}
object is returned, otherwise a \code{\link[data.table]{data.table}} object
is returned.}
}
\value{
a \code{\link[GenomicRanges]{GRanges}} object if \code{is_GRanges}
 is TRUE, otherwise a \code{\link[data.table]{data.table}} object.
}
\description{
\code{read_bs_encode_haib} reads a file containing methylation data from
BS-Seq experiments using the \code{\link{scan}} function.
The BS-Seq file should be in ENCODE HAIB \code{bed} format.
}
\examples{
\donttest{
# Download the files and change the working directory to that location
file <- "name_of_bs_encode_file"
rrbs <- read_bs_encode_haib(file)}

}
\seealso{
\code{\link{read_bs_bismark_cov}},
 \code{\link{read_rna_encode_caltech}}
}

