% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_chrom_size.R
\name{read_chrom_size}
\alias{read_chrom_size}
\title{Read genome chromosome sizes file.}
\usage{
read_chrom_size(file)
}
\arguments{
\item{file}{The name of the file to read data values from.}
}
\value{
A \code{\link[data.table]{data.table}} object.
}
\description{
\code{read_chrom_size} reads a file containing genome chromosome sizes using
 the \code{\link[data.table]{fread}} function.
}
\examples{
chr_file <- system.file("extdata", "hg19.chr.sizes", package = "BPRMeth")
chr_data <- read_chrom_size(chr_file)

# Extract the size of the chr1
chr_data[1]

}
\author{
C.A.Kapourani \email{C.A.Kapourani@ed.ac.uk}
}
\seealso{
\code{\link{read_rna_encode_caltech}},
 \code{\link{read_bs_encode_haib}}
}

