% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_encode_cgi.R
\name{read_encode_cgi}
\alias{read_encode_cgi}
\title{Read file containing CpG island locations}
\usage{
read_encode_cgi(file, is_GRanges = TRUE)
}
\arguments{
\item{file}{The name of the file to read data values from.}

\item{is_GRanges}{Logical: if TRUE a GRanges object is returned, otherwise
a data.frame object is returned.}
}
\value{
A \code{\link[data.table]{data.table}} object.
}
\description{
\code{read_encode_cgi} reads a file containing CpG island (CGI) locations
 in the human genome using the \code{\link[data.table]{fread}} function.
}
\seealso{
\code{\link{read_rna_encode_caltech}},
 \code{\link{read_bs_encode_haib}}
}

