% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{read_expr}
\alias{read_expr}
\title{Read expression data file}
\usage{
read_expr(file, log2_transf = FALSE)
}
\arguments{
\item{file}{File name}

\item{log2_transf}{Logical, whether to log2 transform the expression data.}
}
\value{
A \code{\link[data.table]{data.table}} object.
}
\description{
\code{read_expr} reads a file containing expression data. It
  should contain two columns: "id", "expr" and columns should be in tab
  delimited format.
}
\examples{
\dontrun{
# Download the files and change the working directory to that location
file <- "name_of_expr_file"
anno_dt <- read_expr(file)

# Extract feature id
expr_ids <- anno_dt$id
}

}
\seealso{
\code{\link{read_met}}, \code{\link{read_anno}}
}
\author{
C.A.Kapourani \email{C.A.Kapourani@ed.ac.uk}
}
