% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_model_gex.R
\name{train_model_gex}
\alias{train_model_gex}
\title{Train gene expression model from methylation profiles}
\usage{
train_model_gex(formula = NULL, model_name = "svm", train,
  is_summary = TRUE)
}
\arguments{
\item{formula}{An object of class \code{\link[stats]{formula}}, e.g. see
\code{\link[stats]{lm}} function. If NULL, the simple linear regression
model is used.}

\item{model_name}{A string denoting the regression model. Currently,
available models are: \code{"svm"}, \code{"randomForest"}, \code{"rlm"},
\code{"mars"} and \code{"lm"}.}

\item{train}{The training data.}

\item{is_summary}{Logical, print the summary statistics.}
}
\value{
A list containing the following elements: \itemize{ \item
  \code{formula}: The formula that was used. \item \code{gex_model}: The
  fitted model. \item \code{train_pred} The predicted values for the training
  data. \item \code{train_errors}: The training error metrics. }
}
\description{
\code{train_model_gex} trains a regression model for predicting gene
expression levels by taking as input the higher order methylation features
extracted from specific genomic regions.
}
\examples{
# Create synthetic data
train_data <- data.frame(x = rnorm(20), y=rnorm(20, 1, 4))
res <- train_model_gex(formula = y~., train = train_data)

# Using a different model
res <- train_model_gex(model_name = "randomForest", train = train_data)

}
\author{
C.A.Kapourani \email{C.A.Kapourani@ed.ac.uk}
}
\seealso{
\code{\link{predict_model_gex}}
}

